Automatic detection of p-value
For "Mean method", the automatic detection of pvalue is added. On the curve representing the percentage of different regions within the genome as a function of the p-value used for the statistical (differential) test, the point furthest from the line connecting the two ends of the curve is the optimal p-value (= detection of the curvature of a graph).